+++ Modules/Intent.Modules.Modelers.UI.Core/release-notes.md
@@ -4,3 +4,4 @@
 - Improvement : Support for mapping `Menu Item`s which are children of `Menu Item`s.
 - Improvement : Add `On Selected` event for `Select` components.
 - Improvement : Support for `Radio Button Group`
+- Improvement : Support for `Disabled` binding on The `Button` component.